rap_generator/requirements.txt
functions-framework
flask
openai==1.29.0
anthropic==0.25.7
google-generativeai==0.5.2
tiktoken==0.7.0
langchain==0.1.16
langchain-community==0.0.34
langchain-openai==0.1.3
langchain-core==0.1.46
langchain-google-genai==1.0.3
langchain-anthropic==0.1.11
python-dotenv==1.0.0
httpx == 0.27.2
